# perf: dynamic import dev server